Code much more idiomatic

+++ src/decoder.rs
@@ -32,14 +32,14 @@ impl<'a> Decoder<'a> {
         let android_resources_cursor = Cursor::new(decoder.buffer_android);
         Executor::arsc(android_resources_cursor, &mut decoder.visitor).chain_err(|| "Could not read android lib resources")?;
 
-        let cursor: Cursor<&[u8]> = Cursor::new(decoder.buffer_apk);
+        let cursor = Cursor::new(decoder.buffer_apk);
         Executor::arsc(cursor, &mut decoder.visitor).chain_err(|| "Could not read target APK resources")?;
 
         Ok(decoder)
     }
 
     pub fn get_resources(&self) -> &'a Resources {
-        &self.visitor.get_resources()
+        self.visitor.get_resources()
     }
 
     pub fn as_xml(&self, content: &[u8]) -> Result<String> {
@@ -52,18 +52,17 @@ impl<'a> Decoder<'a> {
             Some(ref root) => {
                 match *visitor.get_string_table() {
                     Some(_) => {
-                        return Xml::encode(
-                            visitor.get_namespaces(),
-                            root,
-                            visitor.get_resources(),
-                            self.get_resources(),
-                        ).chain_err(|| "Could note encode XML");
-                    },
+                        return Xml::encode(visitor.get_namespaces(),
+                                           root,
+                                           visitor.get_resources(),
+                                           self.get_resources())
+                            .chain_err(|| "Could note encode XML");
+                    }
                     None => {
                         println!("No string table found");
                     }
                 }
-            },
+            }
             None => {
                 println!("No root on target XML");
             }
@@ -95,32 +94,33 @@ impl<'a> Apk<'a> {
         Ok(apk)
     }
 
-    /// It exports to target output_path the contents of the APK, transcoding the binary XML files found on it.
+    /// It exports to target output_path the contents of the APK, transcoding the binary XML files
+    /// found on it.
     pub fn export<P: AsRef<Path>>(&mut self, output_path: P, force: bool) -> Result<()> {
-        match fs::create_dir(&output_path) {
-            Err(_) => {
-                if force {
-                    fs::remove_dir_all(&output_path).chain_err(|| "Could not clean target directory")?;
-                    fs::create_dir(&output_path).chain_err(|| "Error creating the output folder")?;
-                }
-            }
-            _ => (),
+        if fs::create_dir(&output_path).is_err() && force {
+            fs::remove_dir_all(&output_path).chain_err(|| "Could not clean target directory")?;
+            fs::create_dir(&output_path).chain_err(|| "Error creating the output folder")?;
         }
 
         // Iterate over all the files on the ZIP and extract them
         for i in 0..self.handler.len() {
             let (file_name, contents) = {
-                let mut current_file = self.handler.by_index(i).chain_err(|| "Could not read ZIP entry")?;
+                let mut current_file =
+                    self.handler.by_index(i).chain_err(|| "Could not read ZIP entry")?;
                 let mut contents = Vec::new();
-                current_file.read_to_end(&mut contents).chain_err(|| format!("Could not read: {}", current_file.name()))?;
+                current_file.read_to_end(&mut contents)
+                    .chain_err(|| format!("Could not read: {}", current_file.name()))?;
                 let is_xml = current_file.name().to_string();
 
                 (is_xml, contents)
             };
 
-            let contents = if (file_name.starts_with("res/") && file_name.ends_with(".xml")) || file_name == "AndroidManifest.xml" {
+            let contents = if (file_name.starts_with("res/") && file_name.ends_with(".xml")) ||
+                              file_name == "AndroidManifest.xml" {
                 let new_content = contents.clone();
-                let out = self.decoder.as_xml(&new_content).chain_err(|| format!("Could not decode: {}", file_name))?;
+                let out = self.decoder
+                    .as_xml(&new_content)
+                    .chain_err(|| format!("Could not decode: {}", file_name))?;
 
                 out.into_bytes()
             } else {
@@ -133,7 +133,10 @@ impl<'a> Apk<'a> {
         Ok(())
     }
 
-    fn write_file<B: AsRef<Path>, R: AsRef<Path>>(base_path: B, relative: R, content: &[u8]) -> Result<()> {
+    fn write_file<B: AsRef<Path>, R: AsRef<Path>>(base_path: B,
+                                                  relative: R,
+                                                  content: &[u8])
+                                                  -> Result<()> {
         let full_path = base_path.as_ref().join(&relative);
         // println!("Full path: {}", full_path.display());
         fs::create_dir_all(full_path.parent().unwrap()).chain_err(|| "Could not create the output dir")?;
@@ -143,7 +146,7 @@ impl<'a> Apk<'a> {
             .open(full_path)
             .chain_err(|| "Could not open file to write")?;
 
-        descriptor.write_all(&content).chain_err(|| "Could not write to target file")?;
+        descriptor.write_all(content).chain_err(|| "Could not write to target file")?;
         descriptor.sync_all().chain_err(|| "Could not flush")?;
 
         Ok(())
